attack 1 check admin role

+++ src/routes/franchiseRouter.js
@@ -98,6 +98,9 @@ franchiseRouter.delete(
   '/:franchiseId',
   asyncHandler(async (req, res) => {
     const franchiseId = Number(req.params.franchiseId);
+    if (!req.user.isRole(Role.Admin)) {
+      throw new StatusCodeError('unable to delete a franchise', 403);
+    }
     await DB.deleteFranchise(franchiseId);
     res.json({ message: 'franchise deleted' });
   })
